third_party/blink/web_tests/external/wpt/url/url-setters-stripping.any-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Setting protocol with leading U+0009 (https:)
  assert_equals: property expected "http:" but got "https:"
[FAIL] Setting protocol with U+0009 before inserted colon (https:)
  assert_equals: property expected "http:" but got "https:"
[FAIL] Setting protocol with leading U+000A (https:)
  assert_equals: property expected "http:" but got "https:"
[FAIL] Setting protocol with U+000A before inserted colon (https:)
  assert_equals: property expected "http:" but got "https:"
[FAIL] Setting protocol with leading U+000D (https:)
  assert_equals: property expected "http:" but got "https:"
[FAIL] Setting protocol with U+000D before inserted colon (https:)
  assert_equals: property expected "http:" but got "https:"
[FAIL] Setting protocol with leading U+0009 (wpt++:)
  assert_equals: property expected "wpt--:" but got "wpt++:"
[FAIL] Setting protocol with U+0009 before inserted colon (wpt++:)
  assert_equals: property expected "wpt--:" but got "wpt++:"
[FAIL] Setting protocol with leading U+000A (wpt++:)
  assert_equals: property expected "wpt--:" but got "wpt++:"
[FAIL] Setting protocol with U+000A before inserted colon (wpt++:)
  assert_equals: property expected "wpt--:" but got "wpt++:"
[FAIL] Setting protocol with leading U+000D (wpt++:)
  assert_equals: property expected "wpt--:" but got "wpt++:"
[FAIL] Setting protocol with U+000D before inserted colon (wpt++:)
  assert_equals: property expected "wpt--:" but got "wpt++:"
Harness: the test ran to completion.
